test(lm): fold Responses contract and dialect probes into existing test files
Move the request-contract tests into test_lm.py beside the existing
Responses coverage (reusing make_response) and the live dialect probes
into test_lm_direct_live.py (same llm_call marker and env-skip pattern),
instead of adding two new test files.
