+++ tests/clients/test_lm.py
@@ -1895,3 +1895,245 @@ async def test_streaming_passes_headers_correctly():
             mock_acompletion.assert_called_once()
             call_kwargs = mock_acompletion.call_args.kwargs
             assert call_kwargs["headers"]["Authorization"] == "Bearer my-custom-token"
+
+
+# ---------------------------------------------------------------------------
+# Responses API request contract: the shapes DSPy must emit for tools,
+# tool_choice, messages, and config on model_type="responses".
+# ---------------------------------------------------------------------------
+
+
+def _chat_shaped_weather_tool():
+    return {
+        "type": "function",
+        "function": {
+            "name": "get_weather",
+            "description": "Get weather.",
+            "parameters": {
+                "type": "object",
+                "properties": {"city": {"type": "string"}},
+                "required": ["city"],
+            },
+        },
+    }
+
+
+class ContractSchema(pydantic.BaseModel):
+    answer: str
+
+
+def test_openai_format_responses_request_maps_tools_choices_and_config():
+    from dspy.clients.openai_format import to_openai_responses_request
+    from dspy.core.types import LMRequest
+
+    request = LMRequest.from_call(
+        model="openai/gpt-5-mini",
+        messages=[{"role": "user", "content": "What is the weather in Paris?"}],
+        tools=[_chat_shaped_weather_tool()],
+        tool_choice={"type": "function", "function": {"name": "get_weather"}},
+        parallel_tool_calls=False,
+        reasoning={"effort": "low", "summary": "auto"},
+        response_format=ContractSchema,
+        max_tokens=123,
+    )
+
+    data = to_openai_responses_request(request)
+
+    assert data["input"] == [
+        {
+            "role": "user",
+            "content": [{"type": "input_text", "text": "What is the weather in Paris?"}],
+        }
+    ]
+    assert data["tools"] == [
+        {
+            "type": "function",
+            "name": "get_weather",
+            "description": "Get weather.",
+            "parameters": {
+                "type": "object",
+                "properties": {"city": {"type": "string"}},
+                "required": ["city"],
+            },
+        }
+    ]
+    assert data["tool_choice"] == {"type": "function", "name": "get_weather"}
+    assert data["parallel_tool_calls"] is False
+    assert data["reasoning"] == {"effort": "low", "summary": "auto"}
+    assert data["max_output_tokens"] == 123
+    # Closed schema, not a raw model_json_schema(): the Responses API rejects
+    # schemas without explicit additionalProperties: false.
+    assert data["text"]["format"] == {
+        "type": "json_schema",
+        "name": "ContractSchema",
+        "schema": {**ContractSchema.model_json_schema(), "additionalProperties": False},
+    }
+
+
+def test_openai_format_responses_request_preserves_unspecified_reasoning_summary():
+    from dspy.clients.openai_format import to_openai_responses_request
+    from dspy.core.types import LMRequest
+
+    request = LMRequest.from_call(
+        model="openai/gpt-5-mini",
+        messages=[{"role": "user", "content": "Say OK."}],
+        reasoning={"effort": "low"},
+    )
+
+    assert request.config.reasoning.summary is None
+    assert to_openai_responses_request(request)["reasoning"] == {"effort": "low"}
+
+
+def test_responses_request_converts_assistant_tool_calls_and_tool_results():
+    from dspy.clients.openai_format import to_openai_responses_request
+    from dspy.core.types import LMRequest
+
+    request = LMRequest.from_call(
+        model="openai/gpt-5-mini",
+        messages=[
+            {"role": "user", "content": "What is the weather?"},
+            {
+                "role": "assistant",
+                "content": None,
+                "tool_calls": [
+                    {
+                        "id": "call_1",
+                        "type": "function",
+                        "function": {"name": "get_weather", "arguments": json.dumps({"city": "Paris"})},
+                    }
+                ],
+            },
+            {"role": "tool", "tool_call_id": "call_1", "name": "get_weather", "content": "sunny"},
+        ],
+    )
+
+    data = to_openai_responses_request(request)
+
+    assert data["input"] == [
+        {"role": "user", "content": [{"type": "input_text", "text": "What is the weather?"}]},
+        {
+            "type": "function_call",
+            "name": "get_weather",
+            "arguments": json.dumps({"city": "Paris"}),
+            "call_id": "call_1",
+        },
+        {"type": "function_call_output", "output": "sunny", "call_id": "call_1"},
+    ]
+
+
+def test_lm_responses_direct_native_tool_calling_uses_responses_tool_shape():
+    response = make_response(
+        [
+            {
+                "type": "function_call",
+                "name": "get_weather",
+                "arguments": json.dumps({"city": "Paris"}),
+                "call_id": "call_1",
+                "id": "fc_1",
+                "status": "completed",
+            }
+        ]
+    )
+
+    with mock.patch("litellm.responses", autospec=True, return_value=response) as responses:
+        lm = dspy.LM("openai/dspy-test-model", model_type="responses", cache=False)
+        outputs = lm("What is the weather in Paris?", tools=[_chat_shaped_weather_tool()])
+
+    # The legacy Responses output shape is unchanged: raw function_call dumps.
+    assert outputs[0]["tool_calls"][0]["name"] == "get_weather"
+    assert outputs[0]["tool_calls"][0]["arguments"] == json.dumps({"city": "Paris"})
+    assert outputs[0]["tool_calls"][0]["call_id"] == "call_1"
+    assert responses.call_args.kwargs["tools"] == [
+        {
+            "type": "function",
+            "name": "get_weather",
+            "description": "Get weather.",
+            "parameters": {
+                "type": "object",
+                "properties": {"city": {"type": "string"}},
+                "required": ["city"],
+            },
+        }
+    ]
+
+
+def test_lm_responses_passes_hosted_tools_through_unchanged():
+    response = make_response(
+        [{"type": "message", "role": "assistant", "content": [{"type": "output_text", "text": "OK", "annotations": []}], "id": "msg_1", "status": "completed"}]
+    )
+    hosted_tool = {"type": "web_search", "search_context_size": "low"}
+
+    with mock.patch("litellm.responses", autospec=True, return_value=response) as responses:
+        lm = dspy.LM("openai/dspy-test-model", model_type="responses", cache=False)
+        lm("What is in the news?", tools=[_chat_shaped_weather_tool(), hosted_tool])
+
+    sent_tools = responses.call_args.kwargs["tools"]
+    assert sent_tools[0]["name"] == "get_weather"
+    assert sent_tools[1] == hosted_tool
+
+    with mock.patch("litellm.responses", autospec=True, return_value=response) as responses:
+        lm = dspy.LM("openai/dspy-test-model", model_type="responses", cache=False)
+        lm("What is in the news?", tools=[hosted_tool])
+
+    assert responses.call_args.kwargs["tools"] == [hosted_tool]
+
+
+def test_lm_responses_preserves_flat_function_tool_with_strict():
+    response = make_response(
+        [{"type": "message", "role": "assistant", "content": [{"type": "output_text", "text": "OK", "annotations": []}], "id": "msg_1", "status": "completed"}]
+    )
+    flat_tool = {
+        "type": "function",
+        "name": "get_weather",
+        "description": "Get weather.",
+        "parameters": {"type": "object", "properties": {"city": {"type": "string"}}, "required": ["city"]},
+        "strict": True,
+    }
+
+    with mock.patch("litellm.responses", autospec=True, return_value=response) as responses:
+        lm = dspy.LM("openai/dspy-test-model", model_type="responses", cache=False)
+        lm("What is the weather in Paris?", tools=[flat_tool])
+
+    assert responses.call_args.kwargs["tools"] == [flat_tool]
+
+
+def test_lm_responses_hoists_nested_strict_to_flat_shape():
+    from dspy.clients.openai_format import to_openai_responses_request
+    from dspy.core.types import LMRequest
+
+    tool = _chat_shaped_weather_tool()
+    tool["function"]["strict"] = True
+    request = LMRequest.from_call(
+        model="openai/gpt-5-mini",
+        messages=[{"role": "user", "content": "What is the weather in Paris?"}],
+        tools=[tool],
+    )
+
+    data = to_openai_responses_request(request)
+
+    assert data["tools"][0]["strict"] is True
+    assert "function" not in data["tools"][0]
+
+
+@pytest.mark.asyncio
+async def test_lm_responses_async_native_tool_calling_matches_sync_contract():
+    response = make_response(
+        [
+            {
+                "type": "function_call",
+                "name": "get_weather",
+                "arguments": json.dumps({"city": "Paris"}),
+                "call_id": "call_1",
+                "id": "fc_1",
+                "status": "completed",
+            }
+        ]
+    )
+
+    with mock.patch("litellm.aresponses", autospec=True, return_value=response) as responses:
+        lm = dspy.LM("openai/dspy-test-model", model_type="responses", cache=False)
+        outputs = await lm.acall("What is the weather in Paris?", tools=[_chat_shaped_weather_tool()])
+
+    assert outputs[0]["tool_calls"][0]["name"] == "get_weather"
+    assert responses.call_args.kwargs["tools"][0]["name"] == "get_weather"
+    assert "function" not in responses.call_args.kwargs["tools"][0]
