+++ tests/clients/test_lm_direct_live.py
@@ -8,8 +8,10 @@ They are intentionally flat rather than parametrized so each test can be run ind
 workflow. Each test skips unless the required provider credential is available.
 """
 
+import json
 import os
 
+import pydantic
 import pytest
 
 import dspy
@@ -309,3 +311,217 @@ def test_live_gemini_chat_direct_reuse_lm_response_as_assistant_turn():
 
     assert "dspy" in _text(first).lower()
     assert "dspy" in _text(follow_up).lower()
+
+
+# ---------------------------------------------------------------------------
+# Responses dialect probes.
+#
+# The Responses request dialect is enforced only by OpenAI's server-side
+# validator: the SDK's request types accept shapes the server rejects, and
+# mocked tests encode the mapper's own beliefs. One minimal real-API request
+# per mapper-supported shape catches the #9943/#9652 class before release.
+# Each probe asserts acceptance plus the smallest semantic check; round-trip
+# probes cover what acceptance alone can't (e.g. referenceable tool-call ids).
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture
+def responses_lm():
+    _require_env("OPENAI_API_KEY")
+    return dspy.LM(
+        os.getenv("LM_FOR_TEST_RESPONSES", "openai/gpt-5-nano"),
+        model_type="responses",
+        cache=False,
+        temperature=1.0,
+        max_tokens=16000,
+    )
+
+
+WEATHER_TOOL_CHAT = {
+    "type": "function",
+    "function": {
+        "name": "get_weather",
+        "description": "Get the current weather for a city.",
+        "parameters": {
+            "type": "object",
+            "properties": {"city": {"type": "string"}},
+            "required": ["city"],
+        },
+    },
+}
+
+# 1x1 transparent PNG.
+TINY_PNG_URI = (
+    "data:image/png;base64,"
+    "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mP8z8BQDwAEhQGAhKmMIQAAAABJRU5ErkJggg=="
+)
+
+
+def _first_tool_call(outputs):
+    out = outputs[0]
+    assert isinstance(out, dict) and out.get("tool_calls"), f"expected a tool call, got: {out!r}"
+    call = out["tool_calls"][0]
+    name = call.get("name") or call.get("function", {}).get("name")
+    arguments = call.get("arguments") or call.get("function", {}).get("arguments")
+    call_id = call.get("call_id") or call.get("id")
+    return name, json.loads(arguments), call_id
+
+
+@pytest.mark.llm_call
+def test_probe_history_all_roles_string_content(responses_lm):
+    outputs = responses_lm(
+        messages=[
+            {"role": "system", "content": "Answer with one word."},
+            {"role": "user", "content": "Say apple."},
+            {"role": "assistant", "content": "apple"},
+            {"role": "user", "content": "Now say banana."},
+        ]
+    )
+    assert outputs
+
+
+@pytest.mark.llm_call
+def test_probe_history_list_form_text_content(responses_lm):
+    outputs = responses_lm(
+        messages=[
+            {"role": "user", "content": [{"type": "text", "text": "Say apple."}]},
+            {"role": "assistant", "content": [{"type": "text", "text": "apple"}]},
+            {"role": "user", "content": [{"type": "text", "text": "Now say banana."}]},
+        ]
+    )
+    assert outputs
+
+
+@pytest.mark.llm_call
+def test_probe_developer_role(responses_lm):
+    outputs = responses_lm(
+        messages=[
+            {"role": "developer", "content": "Answer with one word."},
+            {"role": "user", "content": "Say apple."},
+        ]
+    )
+    assert outputs
+
+
+@pytest.mark.llm_call
+def test_probe_image_content(responses_lm):
+    outputs = responses_lm(
+        messages=[
+            {
+                "role": "user",
+                "content": [
+                    {"type": "text", "text": "One word: what color dominates this image?"},
+                    {"type": "image_url", "image_url": {"url": TINY_PNG_URI}},
+                ],
+            }
+        ]
+    )
+    assert outputs
+
+
+@pytest.mark.llm_call
+def test_probe_nested_chat_tools_return_a_call(responses_lm):
+    outputs = responses_lm("What is the weather in Paris? Use the tool.", tools=[WEATHER_TOOL_CHAT])
+    name, args, call_id = _first_tool_call(outputs)
+    assert name == "get_weather"
+    assert args.get("city") == "Paris"
+    assert call_id
+
+
+@pytest.mark.llm_call
+def test_probe_flat_tool_with_strict(responses_lm):
+    flat_strict = {
+        "type": "function",
+        "name": "get_weather",
+        "description": "Get the current weather for a city.",
+        "parameters": {
+            "type": "object",
+            "properties": {"city": {"type": "string"}},
+            "required": ["city"],
+            "additionalProperties": False,
+        },
+        "strict": True,
+    }
+    outputs = responses_lm("What is the weather in Oslo? Use the tool.", tools=[flat_strict])
+    name, args, _ = _first_tool_call(outputs)
+    assert name == "get_weather"
+    assert args == {"city": "Oslo"}
+
+
+@pytest.mark.llm_call
+def test_probe_hosted_web_search_tool(responses_lm):
+    # The pinned openai SDK models only the "web_search_preview" hosted-tool
+    # shape; the newer "web_search" shape fails in litellm's response parsing.
+    outputs = responses_lm(
+        "Search the web: what year is it? Answer briefly.", tools=[{"type": "web_search_preview"}]
+    )
+    assert outputs
+
+
+@pytest.mark.llm_call
+def test_probe_forced_tool_choice_chat_shape(responses_lm):
+    outputs = responses_lm(
+        "Talk about the weather in Lima.",
+        tools=[WEATHER_TOOL_CHAT],
+        tool_choice={"type": "function", "function": {"name": "get_weather"}},
+    )
+    name, _, _ = _first_tool_call(outputs)
+    assert name == "get_weather"
+
+
+@pytest.mark.llm_call
+def test_probe_tool_choice_none_suppresses_calls(responses_lm):
+    outputs = responses_lm(
+        "What is the weather in Lima? Use the tool.",
+        tools=[WEATHER_TOOL_CHAT],
+        tool_choice="none",
+    )
+    out = outputs[0]
+    assert not (isinstance(out, dict) and out.get("tool_calls"))
+
+
+@pytest.mark.llm_call
+def test_probe_response_format_pydantic(responses_lm):
+    class Answer(pydantic.BaseModel):
+        word: str
+
+    outputs = responses_lm("Reply with the word apple.", response_format=Answer)
+    text = outputs[0]["text"] if isinstance(outputs[0], dict) else outputs[0]
+    assert Answer.model_validate_json(text).word
+
+
+@pytest.mark.llm_call
+def test_probe_reasoning_effort(responses_lm):
+    outputs = responses_lm("Say apple.", reasoning_effort="low")
+    assert outputs
+
+
+@pytest.mark.llm_call
+def test_probe_tool_round_trip_ids_are_referenceable(responses_lm):
+    """Acceptance alone can't catch id bugs: a request that sends back the
+    wrong id (the fc_* item id instead of call_id) fails only on this second
+    turn."""
+    outputs = responses_lm("What is the weather in Berlin? Use the tool.", tools=[WEATHER_TOOL_CHAT])
+    name, args, call_id = _first_tool_call(outputs)
+
+    followup = responses_lm(
+        messages=[
+            {"role": "user", "content": "What is the weather in Berlin? Use the tool."},
+            {
+                "role": "assistant",
+                "content": None,
+                "tool_calls": [
+                    {
+                        "id": call_id,
+                        "type": "function",
+                        "function": {"name": name, "arguments": json.dumps(args)},
+                    }
+                ],
+            },
+            {"role": "tool", "tool_call_id": call_id, "name": name, "content": "It is 22C and sunny."},
+        ],
+        tools=[WEATHER_TOOL_CHAT],
+    )
+    final = followup[0]
+    text = final["text"] if isinstance(final, dict) else final
+    assert text and "22" in text
